Connected expectation and graphs class to GUI

+++ Stock_Price_Tracker/Sorted_Into_Classes/Graphs.py
@@ -1,13 +1,18 @@
 import pandas as pd
 import matplotlib.pyplot as plt
 
+
+
+#Class which uses the pandas and matplotlib classes to create and add data
+# to graphs to help user see his/her stock effciency
+
 class DataManager:
     def __init__(self):
         """Initialize an empty DataFrame to store data."""
         self.df = pd.DataFrame(columns=["Standard Deviation", "Expectation"])
         
 
-    def add_data(self, standard_deviation: int, expectation: int):
+    def add_data(self, standard_deviation: float, expectation: float):
         """Adds new data to the DataFrame without plotting immediately."""
         new_data = {"Standard Deviation": standard_deviation, "Expectation": expectation}
         self.df = self.df._append(new_data, ignore_index=True)
